refactor: move Providers component to Providers folder
