+++ src/components/Providers/Providers.tsx
@@ -1,18 +1,18 @@
 'use client'
 import React from 'react'
-import { Provider } from "react-redux"
-import { store } from '../redux/store'
-import { SessionProvider } from 'next-auth/react'
+import { Provider as ReduxProvider } from "react-redux"
+import { store } from '@/redux/store'
+import { SessionProvider } from 'next-auth/react';
 
 interface Props { children: React.ReactNode };
 
 function Providers({ children }: Props) {
     return (
-        <Provider store={store}>
+        <ReduxProvider store={store}>
             <SessionProvider>
                 {children}
             </SessionProvider>
-        </Provider>
+        </ReduxProvider>
     )
 }
